refactor(view): 사다리 출력 메소드 역할 재분배
findConnectionMark: 연결상태에 따른 마크 선택
