+++ src/main/java/ladder/domain/enums/NameFormat.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.enums;
 
 import java.util.Arrays;
 import java.util.Optional;
