+++ src/main/java/ladder/view/OutputView.java
@@ -1,7 +1,7 @@
 package ladder.view;
 
 import ladder.domain.Ladder;
-import ladder.domain.NameFormat;
+import ladder.domain.enums.NameFormat;
 import ladder.domain.People;
 import ladder.domain.RowLine;
 
@@ -9,7 +9,6 @@ import java.util.List;
 import java.util.StringJoiner;
 
 public class OutputView {
-    //TODO String 상수 Enum으로 변경 고민해보기
     private static final String BLANK = " ";
     private static final String VERTICAL_LINE = "|";
     private static final String HORIZONTAL_LINE = "-";
@@ -40,19 +39,16 @@ public class OutputView {
                 BLANK.repeat(INTERVAL_WIDTH - 1) + VERTICAL_LINE);
 
         for (boolean isConnected : rowLine.getConnection()) {
-            appendLineInterval(isConnected, stringBuilder);
+            stringBuilder.append(findConnectionMark(isConnected).repeat(INTERVAL_WIDTH));
             stringBuilder.append(VERTICAL_LINE);
         }
         System.out.println(stringBuilder);
     }
 
-    private static void appendLineInterval(boolean isConnected, StringBuilder stringBuilder) {
+    private static String findConnectionMark(boolean isConnected) {
         if (isConnected) {
-            stringBuilder.append(HORIZONTAL_LINE.repeat(INTERVAL_WIDTH));
+             return HORIZONTAL_LINE;
         }
-        if (!isConnected) {
-            stringBuilder.append(BLANK.repeat(INTERVAL_WIDTH));
-        }
+         return BLANK;
     }
-
 }
